Refector type in usefetch hook
