+++ src/types/hooks/useFetch.ts
@@ -1,0 +1,12 @@
+import { Method } from 'axios';
+
+interface Props<T> {
+  url: string;
+  method: Method;
+  onSuccess?: (data: T) => void | Promise<void>;
+  onFailure?: (e: unknown) => void | Promise<void>;
+  successMessage?: string;
+  errorMessage?: string;
+}
+
+export default Props;
